requirements.txt
tensorflow
numpy
scipy
librosa
sounddevice
pyserial